use error for failure chan with custom error types
